Add test_allocator to extras/test
